For the people that would like to play with the BI
This repository documents the data cleaning, transformation, and visualization process for Bogotá's public debt (May 2025). The final dashboard is available on Looker Studio:
https://lookerstudio.google.com/reporting/6fed48b2-574e-427e-b672-3cb6a47e9d3b

Important note:
All results in this dashboard are presented as totals per company (entity). If you review the original public records, you will notice that debt figures vary by contract. This reflects a crucial reminder: every data analysis is shaped by how the requirements are defined. In this case, the project focused on aggregating total debt per company to meet a specific reporting need. Different questions or requirements will very likely lead to different aggregations and insights.

Please refer to the dashboard and methodology for details or to adapt this pipeline for other analytical perspectives